fix: Position now based on values between 0 - 1

+++ LectorPotral/src/main/java/com/escapedoom/auth/data/dataclasses/models/escaperoom/nodes/Position.java
@@ -15,8 +15,8 @@ import lombok.NoArgsConstructor;
 @Builder
 public class Position {
 
-    private int x;
+    private double x;
 
-    private int y;
+    private double y;
 
 }
